fix cross compilation problems
The endianness_macro relied on calls to struct.pack which uses the host
python interpreter. This only works if the target has the same
endianness than the host (which is seldom the case). I fixed this
problem by using the AC_C_BIGENDIAN macro in configure.ac. This allows
to get rid of the endianness_macro() in setup.py. The compilation
switches have been adapted accordingly in RIPEMD160.c

Also, cross compilation often needs to pass specific options to
./configure such as --host=linux-mips64 ac_cv_func_malloc_0_nonnull=yes
or --without-gmp

This patch adds the possibility to pass options through the
CONFIGURE_OPTS environment variable.

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>