Adds a static replacement function via an include file (src/rpl_malloc.h) in src/block_template.c and src/stream_template.c
This concludes part 1 of of issue #209

I expect using an ac.get() call in setup.py to check the value of
to add src/rpl_malloc.c to the sources to be compiled
rather than as a static function that is included multiple times.
However, this seems to fit in the style of how src/*_template.c files
are included in the main files.